+++ src/pages/index.tsx
@@ -1,9 +1,7 @@
 import HomePage from '@/modules/HomePage'
+import { getPostQuery, getPosts } from '@/services/api/query'
 import { getCommentIdsQueryKey } from '@/services/subsocial/commentIds'
-import {
-  getPostIdsBySpaceIdQuery,
-  getPostQuery,
-} from '@/services/subsocial/posts'
+import { getPostIdsBySpaceIdQuery } from '@/services/subsocial/posts'
 import { getSubsocialApi } from '@/subsocial-query/subsocial'
 import { getSpaceId } from '@/utils/env/client'
 import { getCommonStaticProps } from '@/utils/page'
@@ -22,13 +20,13 @@ export const getStaticProps = getCommonStaticProps<{
     const promises = postIds.map((postId) => {
       return subsocialApi.blockchain.getReplyIdsByPostId(postId)
     })
-    const postsPromise = subsocialApi.findPublicPosts(postIds)
+    const postsPromise = getPosts(postIds)
 
     const commentIdsByPostId = await Promise.all(promises)
     const posts = await postsPromise
 
     const lastPostIds = commentIdsByPostId.map((ids) => ids[ids.length - 1])
-    const lastPosts = await subsocialApi.findPublicPosts(lastPostIds)
+    const lastPosts = await getPosts(lastPostIds)
 
     getPostIdsBySpaceIdQuery.setQueryData(queryClient, spaceId, {
       spaceId,
