+++ src/services/subsocial/commentIds/optimistic.ts
@@ -1,6 +1,6 @@
+import { getPostQuery } from '@/services/api/query'
 import { PostData } from '@subsocial/api/types'
 import { QueryClient } from '@tanstack/react-query'
-import { getPostQuery } from '../posts'
 import { getCommentIdsQueryKey } from './query'
 import { SendMessageParams } from './types'
